SQL-507: Make MongoDatabaseMetaData abstract and add subtypes
