+++ src/main/java/com/mongodb/jdbc/MongoSQLDatabaseMetaData.java
@@ -1,0 +1,125 @@
+package com.mongodb.jdbc;
+
+import org.bson.BsonInt32;
+import org.bson.BsonString;
+
+import java.sql.*;
+import java.util.ArrayList;
+import java.util.regex.Pattern;
+
+public class MongoSQLDatabaseMetaData extends MongoDatabaseMetaData implements DatabaseMetaData {
+
+    public MongoSQLDatabaseMetaData(MongoConnection conn) {
+        super(conn);
+    }
+
+    @Override
+    public String getSQLKeywords() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getNumericFunctions() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getStringFunctions() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getSystemFunctions() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getTimeDateFunctions() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getTables(
+            String catalog, String schemaPattern, String tableNamePattern, String types[])
+            throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getSchemas() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getCatalogs() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getColumns(
+            String catalog, String schemaPattern, String tableNamePattern, String columnNamePattern)
+            throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getColumnPrivileges(
+            String catalog, String schemaPattern, String tableNamePattern, String columnNamePattern)
+            throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getTablePrivileges(
+            String catalog, String schemaPattern, String tableNamePattern) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getBestRowIdentifier(
+            String catalog, String schema, String table, int scope, boolean nullable)
+            throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getPrimaryKeys(String catalog, String schema, String table)
+            throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getIndexInfo(
+            String catalog, String schema, String table, boolean unique, boolean approximate)
+            throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    //------------------------- JDBC 4.0 -----------------------------------
+
+    @Override
+    public ResultSet getSchemas(String catalog, String schemaPattern) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getClientInfoProperties() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getFunctions(String catalog, String schemaPattern, String functionNamePattern)
+            throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public ResultSet getFunctionColumns(
+            String catalog,
+            String schemaPattern,
+            String functionNamePattern,
+            String columnNamePattern)
+            throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+}
